Refactor : WebSocket Refactoring
