+++ src/main/java/com/sdy/bbbb/entity/Room.java
@@ -1,8 +1,8 @@
-package com.sdy.bbbb.webSocket;
+package com.sdy.bbbb.entity;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.sdy.bbbb.entity.Account;
+import com.sdy.bbbb.entity.Chat;
 import com.sdy.bbbb.entity.TimeStamped;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
