+++ src/main/java/com/sdy/bbbb/repository/ChatRepository.java
@@ -1,5 +1,6 @@
-package com.sdy.bbbb.webSocket;
+package com.sdy.bbbb.repository;
 
+import com.sdy.bbbb.entity.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRepository extends JpaRepository<Chat, Long> {
