+++ src/main/java/com/sdy/bbbb/service/RoomService.java
@@ -1,16 +1,21 @@
-package com.sdy.bbbb.webSocket;
+package com.sdy.bbbb.service;
 
+import com.sdy.bbbb.dto.response.ChatResponseDto;
 import com.sdy.bbbb.dto.response.GlobalResponseDto;
+import com.sdy.bbbb.dto.response.RoomListResponseDto;
+import com.sdy.bbbb.dto.response.RoomResponseDto;
 import com.sdy.bbbb.entity.Account;
+import com.sdy.bbbb.entity.Chat;
+import com.sdy.bbbb.entity.Room;
 import com.sdy.bbbb.exception.CustomException;
 import com.sdy.bbbb.exception.ErrorCode;
 import com.sdy.bbbb.repository.AccountRepository;
+import com.sdy.bbbb.repository.RoomRepository;
 import com.sdy.bbbb.util.Chrono;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
